requirements.txt
autopep8==1.2.4
beautifulsoup4==4.5.1
certifi==2017.1.23
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.9.7
django-allauth==0.25.2
django-ckeditor==5.1.0
django-classy-tags==0.8.0
django-embed-video==1.1.0
django-filebrowser==3.7.2
django-filebrowser-no-grappelli==3.6.2
django-grappelli==2.8.1
django-haystack==2.6.0
django-haystackbrowser==0.6.1
django-registration==2.1.2
django-tinymce==2.3.0
Django-tinymce-filebrowser==0.2.1
elasticsearch==1.9.0
emencia-django-slideshows==1.0.1
groupdocs-python3==2.6.0
gunicorn==19.2.1
httpsig==1.1.2
ipython==2.4.1
oauthlib==2.0.0
PasteDeploy==1.5.2
pep8==1.7.0
Pillow==2.7.0
psycopg2==2.6.2
pycrypto==2.6.1
pyelasticsearch==0.5
PyJWT==1.3.0
pyramid==1.7.3
pyreadline==2.1
python3-openid==3.0.10
pytz==2014.10
repoze.lru==0.6
requests==2.13.0
requests-oauthlib==0.6.2
rollbar==0.13.3
simplejson==3.10.0
six==1.9.0
sorl-thumbnail==12.3
South==0.8.1
stream-django==1.3.1
stream-python==2.3.8
translationstring==1.3
urllib3==1.20
venusian==1.0
virtualenv==15.0.3
WebOb==1.6.1
whitenoise==1.0.6
Whoosh==2.7.4
zope.deprecation==4.1.2
zope.interface==4.3.2
